done FamilyMember

+++ database/migrations/2025_09_24_023752_create_family_members_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->foreign('user_id')->references('id')->on('users');
 
             $table->string('profile_picture')->nullable();
-            $table->integer('identity_number')->unique();
+            $table->bigInteger('identity_number')->unique();
             $table->enum('gender', ['male', 'female']);
             $table->date('date_of_birth');
             $table->string('phone_number')->nullable();
